Handles scrollbar removal without changing document width
